// src/app/models/motorist.model.ts
export class Motorist {

    constructor() {}

    username: string;
    password: string;
    family_name: string;
    given_name: string;
    email_address: string;
    contact_number: string;
    address: string;
    vehicle_information_vehicle_type: string;
    vehicle_information_model_year: string
}